[v0.89.1][maintainability] Split cli/pr_cmd.rs finish responsibility (Closes #1961)
